refactor(import): idIntegrationImporter をドメイン別5モジュールへ分割
1585行の idIntegrationImporter.ts から、エンティティ別の挿入ロジックを
責務ごとに5モジュールへ分離し、オーケストレータ本体を344行へ縮小した。

- importExamCore: Exam根・ExamPage・CropRegion・UserExam・ExamStudent 等の骨格
- importSubtotals: Subtotal・CropSubtotal
- importExamAttachments: 採点マーク/出力設定・OMR・複合解答構造・タグ・学級関連
- importScoring: QuestionScore・ScoreDecision・CompoundAnswerScore
- importSyncRecords: DeletedRecord・DrawingAnnotation・Membership

型規約遵守:
- ローカル重複の type Tx を廃し、共有型 PrismaTransaction(./types)へ統一
- 再エクスポート禁止に従い、processMemberships の外部利用者
  (student-archive/index.ts)の import を新パスへ直接更新
- バレルは変更せず(基準未満のため)。executeIdIntegrationImport は従来どおり参照可能

ロジックは無変更の純粋なコード移動。typecheck 両ターゲット・import-export
252テスト全パス。コードレビュー(high)指摘なし。
